added convenience function getItemlistForCurrentPart();
when navigating through parts (say notes in the PR editor) there can be
multiple Parts that have populated the canvass's itemlist, but we need
to operate with the items that belong to the active Part!

With this function you'll get a list of canvas items for the current
part if there is any.